decode: widen RawAttribute::unique_id to u32 and bounds-check face count
`RawAttribute::unique_id` was `u8` even though the bitstream's `unique_id`
field is `u32` — Draco stores it as a single byte today but the
binding type is `u32`, so future bitstream versions or non-default
encoders could produce values that silently truncated. Match the
binding type and drop the `u8` -> `u32` cast at the consumer side
(splice helper).

`build_raw` now errors with `FaceCountOverflow` if the connectivity
decoder produced more faces than fit in u32 corners (which the flat
indices buffer caps at). Previously the cast `num_corners as u32`
would silently wrap.

+++ draco-oxide/src/io/gltf/draco_decoder.rs
@@ -226,7 +226,7 @@ pub fn splice_glb_remove_draco(input: &[u8]) -> Result<Vec<u8>, Error> {
             for attr in &raw.attributes {
                 let semantic_owned: Option<String> = attr.gltf_semantic.map(|s| s.to_string());
                 let gltf_name = match by_unique_id
-                    .get(&(attr.unique_id as u32))
+                    .get(&attr.unique_id)
                     .cloned()
                     .or(semantic_owned)
                 {
